Code improvements
Morphed FormatOps class into namespace.
Splitted FormatOps into sub-modules.
Added checks on the logs format for missing important fields.
