+++ logdoctor/modules/craplog/modules/CMakeLists.txt
@@ -9,4 +9,5 @@ target_sources(LogDoctor PRIVATE
     ${CMAKE_CURRENT_SOURCE_DIR}/hasher.cpp
 )
 
+add_subdirectory(formats)
 add_subdirectory(hasher)
