+++ logdoctor/modules/craplog/modules/formats/CMakeLists.txt
@@ -1,0 +1,9 @@
+
+target_sources(LogDoctor PRIVATE
+    ${CMAKE_CURRENT_SOURCE_DIR}/check.h
+    ${CMAKE_CURRENT_SOURCE_DIR}/check.cpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/parse.h
+    ${CMAKE_CURRENT_SOURCE_DIR}/parse.cpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/samples.h
+    ${CMAKE_CURRENT_SOURCE_DIR}/samples.cpp
+)
