+++ logdoctor/modules/craplog/modules/formats/check.h
@@ -1,0 +1,16 @@
+#ifndef LOGDOCTOR__CRAPLOG__MODULES__FORMATS__CHECK_H
+#define LOGDOCTOR__CRAPLOG__MODULES__FORMATS__CHECK_H
+
+
+#include "modules/craplog/modules/lib.h"
+
+
+namespace FormatOps::Private
+{
+
+LogsFormat checkFormat( const LogsFormat format );
+
+} // namespace FormatOps::Private
+
+
+#endif // LOGDOCTOR__CRAPLOG__MODULES__FORMATS__CHECK_H
